chore: move operations to client dir
