+++ src/app/(main)/(admin)/sources/page.tsx
@@ -6,7 +6,7 @@ import { DataTableRemote } from '@/components/data-table-remote';
 import { ImportSiteDialog } from '@/components/dialogs/import-site-dialog';
 import { Separator } from '@/components/ui/separator';
 import type { Source } from '@/core/repositories/source';
-import { rescheduleDocumentImportTasks, scheduleDocumentImportTasks } from '@/operations/sources';
+import { rescheduleDocumentImportTasks, scheduleDocumentImportTasks } from '@/client/operations/sources';
 import type { CellContext, ColumnDef } from '@tanstack/react-table';
 import { createColumnHelper } from '@tanstack/table-core';
 import { format } from 'date-fns';
