+++ src/client/operations/batchRetry.ts
